+++ webapp/app/landing/views/statisticalFactsheets/ForestAreaPercent/ForestAreaPercent.js
