+++ webapp/app/landing/views/statisticalFactsheets/NaturallyRegeneratingForest/NaturallyRegeneratingForest.js
