+++ webapp/app/landing/views/statisticalFactsheets/NaturallyRegeneratingForest/index.js
