+++ webapp/app/landing/views/statisticalFactsheets/NaturallyRegeneratingForest/naturallyRegeneratingForest.less
