+++ webapp/app/landing/views/statisticalFactsheets/PrimaryDesignatedManagementObjective/index.js
