+++ webapp/app/landing/views/statisticalFactsheets/PrimaryDesignatedManagementObjective/primaryDesignatedManagementObjective.less
