+++ webapp/app/landing/views/statisticalFactsheets/PrimaryForest/PrimaryForest.js
@@ -2,7 +2,7 @@ import './primaryForest.less'
 import React from 'react'
 import PropTypes from 'prop-types'
 import { useI18n } from '@webapp/components/hooks'
-import * as APIUtils from '@webapp/app/statisticalFactsheets/utils/apiUtils'
+import * as APIUtils from '@webapp/app/landing/views/statisticalFactsheets/utils/apiUtils'
 import * as ChartUtils from '../utils/chartUtils'
 import useStatisticalFactsheetsState from '../hooks/useStatisticalFactsheetsState'
 import Chart from '../components/chart'
