+++ webapp/app/landing/views/statisticalFactsheets/components/chart/Chart.js
