+++ webapp/app/landing/views/statisticalFactsheets/components/chart/index.js
