+++ webapp/app/landing/views/statisticalFactsheets/components/levelSelection/getLabel.js
