+++ webapp/app/landing/views/statisticalFactsheets/components/levelSelection/levelSelection.js
