+++ webapp/app/landing/views/statisticalFactsheets/components/table/index.js
