+++ webapp/app/landing/views/statisticalFactsheets/utils/chartUtils.js
