+++ webapp/app/landing/views/statisticalFactsheets/utils/propUtils.js
